% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_file_type.R
\name{detect_file_type}
\alias{detect_file_type}
\title{Detect File Type}
\usage{
detect_file_type(input)
}
\arguments{
\item{input}{the absolute file path of the directory of raw files for conversion}
}
\value{
a character string of the file type for conversion
}
\description{
Detect the main file type in a directory of raw files
}
