% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_convert.R
\name{slurm_convert}
\alias{slurm_convert}
\title{Convert Raw Files using msconvert and SLURM}
\usage{
slurm_convert(input, output, format_out, conversion_args)
}
\arguments{
\item{input}{the absolute file path of the directory of raw files for conversion}

\item{output}{the absolute file path where converted data files will be saved to}

\item{format_out}{a character string of \code{msconvert} format for conversioin.}

\item{conversion_args}{a character string of \code{msconvert} arguments without the \code{--filter} prefix (ie, \verb{peakPicking true 1-})}
}
\description{
Convert raw mass spectrometry files using a combination of singularity, t
}
\examples{
\dontrun{
slurm_convert(input = 'hpc/storage/my_raw_data', output = 'hpc/home/my_converted_data', format_out = 'mzML', conversion_args = c('peakPicking true 1-'))
}
}
